identity_service/requirements.txt
Django==4.2.2
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
djangorestframework==3.14.0
django_filter==23.2
daphne==4.0.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.16
rest_framework_simplejwt==0.0.2
PyJWT==2.8.0
requests==2.31.0
celery==5.3.4